backend/novel_downloader/novel_src/offical_tools/downloader.py
import base64
import gzip
import json
import time
import urllib3
from typing import Any, Dict, List

import requests
from Crypto.Cipher import AES
from Crypto.Random import get_random_bytes
from Crypto.Util.Padding import pad, unpad

from ..base_system.context import GlobalContext
from .get_iid import get_iid
from .get_version_code import GetVersionCode

urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
requests.packages.urllib3.disable_warnings()

# ===== 常量 =====
STATIC_KEY_PARTS = ["ac25", "c67d", "dd8f", "38c1", "b37a", "2348", "828e", "222e"]
API_BASE_URL = "https://api5-normal-sinfonlineb.fqnovel.com/reading"
AID = "1967"
_IID_EXPIRE_SEC = 12 * 60 * 60  # 12 小时


def _ensure_fresh_iid() -> None:
    cfg = GlobalContext.get_config()
    log = GlobalContext.get_logger()
    
    # 如果使用代理模式或不使用官方API，跳过IID刷新
    use_proxy_api = getattr(cfg, "use_proxy_api", False)
    use_official_api = getattr(cfg, "use_official_api", False)
    
    if use_proxy_api or not use_official_api:
        log.debug("代理模式已启用或官方API已禁用，跳过IID刷新")
        return
    
    now_ms = int(time.time() * 1000)
    spawn_ms = int(getattr(cfg, "iid_spawn_time", "0") or "0")
    if not getattr(cfg, "iid", None) or (now_ms - spawn_ms) // 1000 >= _IID_EXPIRE_SEC:
        log.info("iid 已使用超过 12 小时或不存在，自动更换 …")
        cfg.iid = str(get_iid())
        cfg.iid_spawn_time = str(now_ms)
        cfg.save()


def get_static_key() -> str:
    return "".join(STATIC_KEY_PARTS)


class FqCrypto:
    def __init__(self, key_hex: str):
        self.key = bytes.fromhex(key_hex)
        if len(self.key) != 16:
            raise ValueError(f"Key length mismatch: {self.key.hex()}")
        self.mode = AES.MODE_CBC

    def encrypt(self, data: bytes, iv: bytes) -> bytes:
        return AES.new(self.key, self.mode, iv).encrypt(pad(data, AES.block_size))

    def decrypt(self, data: bytes) -> bytes:
        iv, ct = data[:16], data[16:]
        return unpad(AES.new(self.key, self.mode, iv).decrypt(ct), AES.block_size)

    def build_register_content(self, server_device_id: str, str_val: str) -> str:
        if not server_device_id.isdigit() or not str_val.isdigit():
            raise ValueError("server_device_id 和 str_val 必须为纯数字")
        raw = int(server_device_id).to_bytes(8, "little") + int(str_val).to_bytes(
            8, "little"
        )
        iv = get_random_bytes(16)
        return base64.b64encode(iv + self.encrypt(raw, iv)).decode()


class FqVariable:
    def __init__(
        self, install_id: str, server_device_id: str, aid: str, update_version_code: str
    ):
        self.install_id = install_id
        self.server_device_id = server_device_id
        self.aid = aid
        self.update_version_code = update_version_code


class FqReq:
    def __init__(self, fq_var: FqVariable, *, timeout: int = 10):
        self.var = fq_var
        self._timeout = timeout
        self.session = requests.Session()
        self._crypto = None
        self._key_version = None
        self._fetch_register_key()

    def get_contents(self, chapter_ids: str) -> Dict:
        raw = self._batch_fetch(chapter_ids)
        first = next(iter(raw.get("data", {}).values()), {})
        self._ensure_key_version(first.get("key_version"))
        return self._decrypt_contents(raw)

    def _batch_fetch(self, item_ids: str) -> Dict:
        params = {
            "item_ids": item_ids,
            "update_version_code": self.var.update_version_code,
            "aid": self.var.aid,
            "key_register_ts": "0",
        }
        headers = {"Cookie": f"install_id={self.var.install_id}"}
        
        # 构造完整的URL
        from urllib.parse import urlencode
        url = f"{API_BASE_URL}/reader/batch_full/v"
        query_string = urlencode(params)
        full_url = f"{url}?{query_string}"
        
        # 打印curl命令
        log = GlobalContext.get_logger()
        log.info(f"\n========== CURL Command ==========\n")
        log.info(f"curl -X GET \"{full_url}\" \\")
        log.info(f"  -H \"Cookie: install_id={self.var.install_id}\" \\")
        log.info(f"  -H \"User-Agent: Mozilla/5.0\" \\")
        log.info(f"  --insecure")
        log.info(f"\n==================================\n")
        
        r = self.session.get(
            url,
            headers=headers,
            params=params,
            timeout=self._timeout,
            verify=False,
        )
        
        log.info(f"Response Status Code: {r.status_code}")
        log.info(f"Response Content Length: {len(r.content)} bytes")
        log.info(f"Response Headers: {dict(r.headers)}")
        
        if len(r.content) == 0:
            log.error(f"API returned empty content! This usually means:")
            log.error(f"  1. IID is invalid or expired")
            log.error(f"  2. Chapter ID doesn't exist")
            log.error(f"  3. API blocked the request")
            log.error(f"Request URL: {r.url}")
            log.error(f"Request Headers: {r.request.headers}")
            # Return empty data structure instead of crashing
            return {"data": {}}
        
        if r.status_code != 200:
            log.error(f"Response Content: {r.text[:500]}")
        
        r.raise_for_status()
        return r.json()

    def _fetch_register_key(self) -> None:
        log = GlobalContext.get_logger()
        log.info(f"Fetching register key...")
        static_crypto = FqCrypto(get_static_key())
        payload = {
            "content": static_crypto.build_register_content(
                self.var.server_device_id, "0"
            ),
            "keyver": 1,
        }
        r = self.session.post(
            f"{API_BASE_URL}/crypt/registerkey",
            headers={
                "Cookie": f"install_id={self.var.install_id}",
                "Content-Type": "application/json",
            },
            params={"aid": self.var.aid},
            data=json.dumps(payload).encode(),
            timeout=self._timeout,
            verify=False,
        )
        log.info(f"Register key response status: {r.status_code}, content length: {len(r.content)}")
        r.raise_for_status()
        response_data = r.json()
        log.info(f"Register key response keys: {list(response_data.keys())}")
        data = response_data["data"]
        self._key_version = data["keyver"]
        key_hex = static_crypto.decrypt(base64.b64decode(data["key"]))
        self._crypto = FqCrypto(key_hex.hex())
        log.info(f"Register key fetch successful, key_version: {self._key_version}")

    def _ensure_key_version(self, expected):
        if expected is not None and expected != self._key_version:
            self._fetch_register_key()

    def _decrypt_contents(self, res: Dict) -> Dict:
        if not self._crypto:
            raise RuntimeError("register key 尚未初始化")
        for cid, info in res.get("data", {}).items():
            enc = info.get("content")
            if not enc:
                continue
            raw = self._crypto.decrypt(base64.b64decode(enc))
            try:
                info["content"] = gzip.decompress(raw).decode()
            except:
                info["content"] = raw.decode()
        return res

    def close(self):
        self.session.close()


def search_api(book_name: str, offset: int = 0) -> Dict[str, Any]:
    cfg = GlobalContext.get_config()
    log = GlobalContext.get_logger()
    
    # 搜索API不需要动态IID，使用固定cookie即可
    # 代理模式下也可以使用此搜索功能
    use_proxy_api = getattr(cfg, "use_proxy_api", False)
    use_official_api = getattr(cfg, "use_official_api", False)
    
    if use_proxy_api or not use_official_api:
        log.info("代理模式下使用公开搜索API（不需要IID）")
    else:
        # 只有在官方API模式下才刷新IID
        _ensure_fresh_iid()
    
    # 使用固定的install_id进行搜索，不需要动态IID
    headers = {
        "cookie": "install_id=1229734607899353;",
        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36",
    }
    api = "https://api-lf.fanqiesdk.com/api/novel/channel/homepage/search/search/v1/"
    params = {
        "offset": str(offset),
        "aid": "1967",
        "q": book_name,
    }

    response = requests.get(api, params=params, headers=headers, verify=False, timeout=10)

    response.raise_for_status()

    data = response.json().get("data", {})
    ret_data = data.get("ret_data", [])
    if not isinstance(ret_data, list):
        ret_data = []

    next_offset = data.get("offset")
    try:
        next_offset = int(next_offset) if next_offset is not None else None
    except (TypeError, ValueError):
        next_offset = None

    return {
        "results": ret_data,
        "has_more": bool(data.get("has_more")),
        "next_offset": next_offset,
    }


def download_chapter_official(chapter_id: str) -> Dict:
    cfg = GlobalContext.get_config()
    log = GlobalContext.get_logger()
    
    # 如果使用代理模式，不应调用此函数
    use_proxy_api = getattr(cfg, "use_proxy_api", False)
    use_official_api = getattr(cfg, "use_official_api", False)
    
    if use_proxy_api or not use_official_api:
        log.error("代理模式下不支持使用官方章节下载API")
        raise RuntimeError("代理模式下不支持使用官方章节下载API，请使用代理API下载")
    
    _ensure_fresh_iid()
    var = FqVariable(
        install_id=cfg.iid,
        server_device_id=str(int(cfg.iid) - 4096),
        aid=AID,
        update_version_code=GetVersionCode.get(),
    )
    fq = FqReq(var)
    try:
        data = fq.get_contents(chapter_id)
        return data
    finally:
        fq.close()


def spawn_iid() -> None:
    """自动生成 install_id 并验证能否正常取正文"""
    cfg = GlobalContext.get_config()
    log = GlobalContext.get_logger()
    
    # 如果使用代理模式，不应调用此函数
    use_proxy_api = getattr(cfg, "use_proxy_api", False)
    use_official_api = getattr(cfg, "use_official_api", False)
    
    if use_proxy_api or not use_official_api:
        log.info("代理模式已启用或官方API已禁用，无需生成IID")
        return

    _ensure_fresh_iid()  # ← 自动确保有可用 iid
    log.info("当前 iid=%s", cfg.iid)

    fq_var = FqVariable(
        install_id=cfg.iid,
        server_device_id=str(int(cfg.iid) - 4096),
        aid=AID,
        update_version_code=GetVersionCode.get(),
    )
    fq_req = FqReq(fq_var)
    try:
        # 随便请求一章测试
        for attempt in range(1, 6):
            try:
                fq_req.get_contents("7310102404588896783")
                log.info(f"验证成功（第 {attempt} 次）")
                break
            except Exception as e:
                log.warning(f"验证失败（第 {attempt} 次）：{e}")
                time.sleep(0.3)
        else:
            raise RuntimeError("连续验证失败，请检查网络或算法")
    finally:
        fq_req.close()
    log.info("iid 获取并验证成功！")
